// mint-app/src/pages/project-management/enums/meeting-status.enums.ts
export enum MEETING_STATUS_ENUM {
  FINAL = 'FINAL',
  PROPASAL = 'PROPASAL',
  INITIAL = 'INITIAL',
}

export enum BUTTON_CONFIRM_ENUM {
  ACCEPT = 'ACCEPT', // Accept or Reject Suggestion From Opposite
  PROPOSE_NEW = 'PROPOSE_NEW', // Create Grey
  PROPOSE_MEETINGS = 'PROPOSE_MEETINGS', // Create Green
}